requirements.txt

robotframework
python-can
